Add format_url function
This is a generic function that wraps uritemplate, and also adds the
ability to add a hostname. This will be a fairly common pattern in this
library because relative paths will be well-known, but the hostname
needs to be configured.

This is based on gidgethub. I've Added the license for gidgethub
(licenses/gidgethub.txt) along with attribution in the module docstring
to give credit to Gidgethub for the sansio design and basis for this
implementation.