Create a sans-io client for the Schema Registry
This client is based on the sans-io architecture
(https://sans-io.readthedocs.io) and heavily influenced by the Gidgethub
implementation. The idea is that the sansio.RegistryApi client knows how
to format requests and process responses for the Confluence Schema
Registry API, but doesn't include an HTTP client itself. This means it's
easy to subclass RegistryApi with a mock implementation for unit
testing, and with other subclasses to integrate with specific HTTP
clients.

Some things that can be improved from this starting point:

- More tests (I've got cursory tests set up)
- A clear exception hierachy for the different failure odes of the
  schema registry.

This also adds pytest-asyncio, which adds a useful decorator for testing
coroutines.