Add an aiohttp-specific client for RegistryApi
This client subclass integrates with aiohttp itself and is what I expect
most users to actually use. There's not a lot to it given that the
sansio base client has all the logic.

I've added an aiohttp setuptools extra to make it easy to install
aiohttp alongside kafkit, though because the session is created by the
user, there isn't (yet) an import dependency on aiohttp inside kafkit.