Add registry-enabled Serializer + Deserializer
This adds Serializer and Deserializer classes that can be used by
producers and consumers to send and recieve messages in the Confluent
Wire Format (https://docs.confluent.io/current/schema-registry/docs/serializer-formatter.html#wire-format)
This means that schema ID in the schema registry are included as a
prefix to the avro-encoded message.

The Serializer is dedicated to working for a specific schema. This way
is can be used as a value_serializer or key_serializer by the Producer
(see
https://aiokafka.readthedocs.io/en/stable/examples/serialize_and_compress.html#serialization-and-compression).

The Deserializer needs to react to whatever schema ID gets send to it.
For that reason, it can't implement a __call__ (it'd need to be a
coroutine), and thus must always be used externally to the consumer.