Add mypy-based type checking
This adds mypy type checking (configured in tox at the "typing"
environment) and adds type annotations throughout the code base.

In doing this, one of the things I learned is that the technique the
SchemaCache used of returning either the schema ID or schema depending
on what you provide to __getindex__ relies on typing.overload to
accurately describe the return type for each scenario. It'd be better
just to have dedicated method for retrieving either type.