Add RecordNameSchemaManager
The purpose of this manager is to provide an integrated, opinionated,
workflow for working with local schemas that are reflected in the
Confluent Schema Registry.

Currently implements a workflow for registering local schemas with the
schema registry and resetting compatibility, and support for serializing
a message based on the locally maintained schemas.

Test is done with the docker-compose-based Confluent Platform deployment.

This is necessary because we can't use the mock schema registry client
given the complexity of the interactions the manager has with the schema
registry.

This test is mostly a smoke test following the expected usage pattern
for the RecordNameSchemaManager.